+++ packages/angular/CHANGELOG.md
@@ -1,1 +1,8 @@
+## @lunora/angular [1.0.0-alpha.2](https://github.com/anolilab/lunora/compare/%40lunora%2Fangular%401.0.0-alpha.1...%40lunora%2Fangular%401.0.0-alpha.2) (2026-07-04)
+
+
+### Dependencies
+
+* **@lunora/client:** upgraded to 1.0.0-alpha.18
+
 ## @lunora/angular 1.0.0-alpha.1 (2026-07-03)