+++ packages/solid/CHANGELOG.md
@@ -1,3 +1,10 @@
+## @lunora/solid [1.0.0-alpha.20](https://github.com/anolilab/lunora/compare/%40lunora%2Fsolid%401.0.0-alpha.19...%40lunora%2Fsolid%401.0.0-alpha.20) (2026-07-04)
+
+
+### Dependencies
+
+* **@lunora/client:** upgraded to 1.0.0-alpha.18
+
 ## @lunora/solid [1.0.0-alpha.19](https://github.com/anolilab/lunora/compare/%40lunora%2Fsolid%401.0.0-alpha.18...%40lunora%2Fsolid%401.0.0-alpha.19) (2026-07-03)